Removes duplication and just use Playerstub
