+++ src/test/java/com/samhan/GameTest.java
@@ -1,8 +1,7 @@
 package com.samhan;
 
 import com.samhan.Fakes.DisplaySpy;
-import com.samhan.Fakes.PlayerSpy;
-import com.samhan.player.Player;
+import com.samhan.Fakes.PlayerStub;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -13,14 +12,14 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 public class GameTest {
-    private PlayerSpy player1;
-    private PlayerSpy player2;
+    private PlayerStub player1;
+    private PlayerStub player2;
     private DisplaySpy display;
 
     @Before
     public void setUp() {
-        player1 = new PlayerSpy(Marker.X);
-        player2 = new PlayerSpy(Marker.O);
+        player1 = new PlayerStub(Marker.X, new LinkedList<>(Arrays.asList(0, 2, 3, 5, 7)));
+        player2 = new PlayerStub(Marker.O, new LinkedList<>(Arrays.asList(1, 4, 6, 8)));
         display = new DisplaySpy();
     }
 
@@ -91,29 +90,6 @@ public class GameTest {
         assertThat(display.renderTimesCalled, is(10));
         assertThat(player1.nextMoveTimesCalled, is(5));
         assertThat(player2.nextMoveTimesCalled, is(4));
-    }
-
-    private class PlayerStub implements Player {
-        private final Marker marker;
-        private final LinkedList<Integer> moves;
-        public int nextMoveTimesCalled;
-
-        public PlayerStub(Marker marker, LinkedList<Integer> moves) {
-            this.marker = marker;
-            this.moves = moves;
-            this.nextMoveTimesCalled = 0;
-        }
-
-        @Override
-        public Marker getMarker() {
-            return marker;
-        }
-
-        @Override
-        public int nextMove(Board board) {
-            this.nextMoveTimesCalled++;
-            return moves.remove();
-        }
     }
 }
